MET-4588 RevisionResource migrated to new A&A model

+++ service/mcs/rest/src/main/java/eu/europeana/cloud/service/mcs/rest/RevisionResource.java
@@ -1,11 +1,14 @@
 package eu.europeana.cloud.service.mcs.rest;
 
 import com.google.common.collect.Sets;
+import eu.europeana.cloud.common.model.CompoundDataSetId;
+import eu.europeana.cloud.common.model.DataSet;
+import eu.europeana.cloud.common.model.Representation;
 import eu.europeana.cloud.common.model.Revision;
 import eu.europeana.cloud.common.utils.Tags;
 import eu.europeana.cloud.service.mcs.DataSetService;
 import eu.europeana.cloud.service.mcs.RecordService;
-import eu.europeana.cloud.service.mcs.exception.ProviderNotExistsException;
+import eu.europeana.cloud.service.mcs.exception.AccessDeniedOrObjectDoesNotExistException;
 import eu.europeana.cloud.service.mcs.exception.RepresentationNotExistsException;
 import eu.europeana.cloud.service.mcs.exception.RevisionIsNotValidException;
 import eu.europeana.cloud.service.mcs.utils.ParamUtil;
@@ -18,15 +21,18 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.server.ServletServerHttpRequest;
-import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.access.PermissionEvaluator;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContext;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.util.UriComponentsBuilder;
 
 import javax.servlet.http.HttpServletRequest;
 import java.net.URI;
-import java.net.URISyntaxException;
 import java.util.Arrays;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
 
 import static eu.europeana.cloud.service.mcs.RestInterfaceConstants.*;
@@ -40,10 +46,12 @@ public class RevisionResource {
 
     private final RecordService recordService;
     private final DataSetService dataSetService;
+    private final PermissionEvaluator permissionEvaluator;
 
-    public RevisionResource(RecordService recordService, DataSetService dataSetService) {
+    public RevisionResource(RecordService recordService, DataSetService dataSetService, PermissionEvaluator permissionEvaluator) {
         this.recordService = recordService;
         this.dataSetService = dataSetService;
+        this.permissionEvaluator = permissionEvaluator;
     }
 
     /**
@@ -62,8 +70,6 @@ public class RevisionResource {
      * @statuscode 201 object has been created.
      */
     @PostMapping(value = REVISION_ADD_WITH_PROVIDER_TAG)
-    @PreAuthorize("hasPermission(#cloudId.concat('/').concat(#representationName).concat('/').concat(#version),"
-            + " 'eu.europeana.cloud.common.model.Representation', read)")
     public ResponseEntity<String> addRevision(
             HttpServletRequest httpServletRequest,
             @PathVariable final String cloudId,
@@ -71,20 +77,20 @@ public class RevisionResource {
             @PathVariable final String version,
             @PathVariable String revisionName,
             @PathVariable String revisionProviderId,
-            @PathVariable String tag) throws RepresentationNotExistsException, RevisionIsNotValidException {
+            @PathVariable String tag) throws RepresentationNotExistsException, RevisionIsNotValidException, AccessDeniedOrObjectDoesNotExistException {
 
         ParamUtil.validate("tag", tag,
                 Arrays.asList(Tags.ACCEPTANCE.getTag(), Tags.PUBLISHED.getTag(), Tags.DELETED.getTag()));
-
-        Revision revision = new Revision(revisionName, revisionProviderId);
-        setRevisionTags(revision, new HashSet<>(Arrays.asList(tag)));
-        addRevision(cloudId, representationName, version, revision);
-
-        // insert information in extra table
-        recordService.insertRepresentationRevision(cloudId, representationName, revisionProviderId,
-                revisionName, version, revision.getCreationTimeStamp());
-
-        return createResponseEntity(httpServletRequest, null);
+        //
+        Representation representation = buildRepresentationFromRequestParameters(cloudId, representationName, version);
+        Revision revision = buildRevisionFromRequestParams(revisionName, revisionProviderId, tag);
+        //
+        if (isUserAllowedToAddRevisionTo(representation)) {
+            addRevisionToRepresentationVersion(revision, representation);
+            return createResponseEntity(httpServletRequest, null);
+        } else {
+            throw new AccessDeniedOrObjectDoesNotExistException();
+        }
     }
 
     /**
@@ -97,22 +103,22 @@ public class RevisionResource {
      * @statuscode 201 object has been created.
      */
     @PostMapping(value = REVISION_ADD, consumes = {MediaType.APPLICATION_JSON_VALUE})
-    @PreAuthorize("hasPermission(#cloudId.concat('/').concat(#representationName).concat('/').concat(#version),"
-            + " 'eu.europeana.cloud.common.model.Representation', read)")
     public ResponseEntity<?> addRevision(
             HttpServletRequest httpServletRequest,
             @PathVariable final String cloudId,
             @PathVariable final String representationName,
             @PathVariable final String version,
-            @RequestBody Revision revision) throws RevisionIsNotValidException, RepresentationNotExistsException {
+            @RequestBody Revision revision) throws RevisionIsNotValidException, RepresentationNotExistsException, AccessDeniedOrObjectDoesNotExistException {
 
-        addRevision(cloudId, representationName, version, revision);
-
-        // insert information in extra table
-        recordService.insertRepresentationRevision(cloudId, representationName, revision.getRevisionProviderId(),
-                revision.getRevisionName(), version, revision.getCreationTimeStamp());
-
-        return createResponseEntity(httpServletRequest, null);
+        //
+        Representation representation = buildRepresentationFromRequestParameters(cloudId, representationName, version);
+        //
+        if (isUserAllowedToAddRevisionTo(representation)) {
+            addRevisionToRepresentationVersion(revision, representation);
+            return createResponseEntity(httpServletRequest, null);
+        } else {
+            throw new AccessDeniedOrObjectDoesNotExistException();
+        }
     }
 
 
@@ -132,8 +138,6 @@ public class RevisionResource {
      * @statuscode 201 object has been created.
      */
     @PostMapping(value = REVISION_ADD_WITH_PROVIDER)
-    @PreAuthorize("hasPermission(#cloudId.concat('/').concat(#representationName).concat('/').concat(#version),"
-            + " 'eu.europeana.cloud.common.model.Representation', read)")
     public ResponseEntity<Revision> addRevision(
             HttpServletRequest httpServletRequest,
             @PathVariable final String cloudId,
@@ -142,18 +146,19 @@ public class RevisionResource {
             @PathVariable String revisionName,
             @PathVariable String revisionProviderId,
             @RequestParam(defaultValue = "") Set<String> tags ) throws RepresentationNotExistsException,
-                                                        RevisionIsNotValidException {
+            RevisionIsNotValidException, AccessDeniedOrObjectDoesNotExistException {
 
         ParamUtil.validateTags(tags, new HashSet<>(Sets.newHashSet(Tags.ACCEPTANCE.getTag(), Tags.PUBLISHED.getTag(), Tags.DELETED.getTag())));
-        Revision revision = new Revision(revisionName, revisionProviderId);
-        setRevisionTags(revision, tags);
-        addRevision(cloudId, representationName, version, revision);
-
-        // insert information in extra table
-        recordService.insertRepresentationRevision(cloudId, representationName, revisionProviderId, revisionName,
-                version, revision.getCreationTimeStamp());
-
-        return createResponseEntity(httpServletRequest, revision);
+        //
+        Representation representation = buildRepresentationFromRequestParameters(cloudId, representationName, version);
+        Revision revision = buildRevisionFromRequestParams(revisionName, revisionProviderId, tags);
+        //
+        if (isUserAllowedToAddRevisionTo(representation)) {
+            addRevisionToRepresentationVersion(revision, representation);
+            return createResponseEntity(httpServletRequest, revision);
+        } else {
+            throw new AccessDeniedOrObjectDoesNotExistException();
+        }
     }
 
     /**
@@ -165,12 +170,10 @@ public class RevisionResource {
      * @param revisionName       revision name
      * @param revisionProviderId revision provider
      * @param revisionTimestamp  revision timestamp
-     * @throws ProviderNotExistsException
-     * @throws RepresentationNotExistsException
+     * @throws RepresentationNotExistsException when the representation version doesn't exist
+     * @throws AccessDeniedOrObjectDoesNotExistException when user doesn't have privileges for the dataset
      */
     @DeleteMapping(value = REVISION_DELETE)
-    @PreAuthorize("hasPermission(#cloudId.concat('/').concat(#representationName).concat('/').concat(#version),"
-            + " 'eu.europeana.cloud.common.model.Representation', read)")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteRevision(
             @PathVariable String cloudId,
@@ -178,10 +181,69 @@ public class RevisionResource {
             @PathVariable String version,
             @PathVariable String revisionName,
             @PathVariable String revisionProviderId,
-            @RequestParam String revisionTimestamp ) throws RepresentationNotExistsException {
+            @RequestParam String revisionTimestamp ) throws RepresentationNotExistsException, AccessDeniedOrObjectDoesNotExistException {
 
-        DateTime timestamp = new DateTime(revisionTimestamp, DateTimeZone.UTC);
-        dataSetService.deleteRevision(cloudId, representationName, version, revisionName, revisionProviderId, timestamp.toDate());
+        //
+        Representation representation = buildRepresentationFromRequestParameters(cloudId,representationName,version);
+        //
+        if (isUserAllowedToAddRevisionTo(representation)) {
+            DateTime timestamp = new DateTime(revisionTimestamp, DateTimeZone.UTC);
+            dataSetService.deleteRevision(cloudId, representationName, version, revisionName, revisionProviderId, timestamp.toDate());
+        } else {
+            throw new AccessDeniedOrObjectDoesNotExistException();
+        }
+
+    }
+
+    private Representation buildRepresentationFromRequestParameters(String cloudId, String representationName, String version){
+        Representation representation = new Representation();
+        representation.setCloudId(cloudId);
+        representation.setRepresentationName(representationName);
+        representation.setVersion(version);
+        return representation;
+    }
+
+    private Revision buildRevisionFromRequestParams(String revisionName, String revisionProviderId, String tag){
+        Revision revision = new Revision(revisionName, revisionProviderId);
+        setRevisionTags(revision, new HashSet<>(List.of(tag)));
+        return revision;
+    }
+
+    private Revision buildRevisionFromRequestParams(String revisionName, String revisionProviderId, Set<String> tags){
+        Revision revision = new Revision(revisionName, revisionProviderId);
+        setRevisionTags(revision, tags);
+        return revision;
+    }
+
+    private void addRevisionToRepresentationVersion(Revision revision, Representation representation) throws RepresentationNotExistsException, RevisionIsNotValidException {
+        addRevision(
+                representation.getCloudId(),
+                representation.getRepresentationName(),
+                representation.getVersion(),
+                revision);
+
+        // insert information in extra table
+        recordService.insertRepresentationRevision(
+                representation.getCloudId(),
+                representation.getRepresentationName(),
+                revision.getRevisionProviderId(),
+                revision.getRevisionName(),
+                representation.getVersion(),
+                revision.getCreationTimeStamp());
+    }
+
+    private boolean isUserAllowedToAddRevisionTo(Representation representation) throws RepresentationNotExistsException {
+        List<CompoundDataSetId> representationDataSets = dataSetService.getAllDatasetsForRepresentationVersion(representation);
+        if (representationDataSets.size() != 1) {
+            LOGGER.error("Should never happen");
+        } else {
+            SecurityContext ctx = SecurityContextHolder.getContext();
+            Authentication authentication = ctx.getAuthentication();
+            //
+            String targetId = representationDataSets.get(0).getDataSetId() + "/" + representationDataSets.get(0).getDataSetProviderId();
+            return permissionEvaluator.hasPermission(authentication, targetId, DataSet.class.getName(), "read");
+        }
+        return false;
     }
 
     private void addRevision(String globalId, String schema, String version, Revision revision)
@@ -195,9 +257,9 @@ public class RevisionResource {
         dataSetService.updateAllRevisionDatasetsEntries(globalId, schema, version, revision);
     }
 
-    private Revision setRevisionTags(Revision revision, Set<String> tags) {
+    private void setRevisionTags(Revision revision, Set<String> tags) {
         if (tags == null || tags.isEmpty()) {
-            return revision;
+            return;
         }
         if (tags.contains(Tags.ACCEPTANCE.getTag())) {
             revision.setAcceptance(true);
@@ -208,7 +270,6 @@ public class RevisionResource {
         if (tags.contains(Tags.DELETED.getTag())) {
             revision.setDeleted(true);
         }
-        return revision;
     }
 
     private <T> ResponseEntity<T> createResponseEntity(HttpServletRequest httpServletRequest, T entity) {
